Success message display on photo taken. Animation to error and success messages added.
