+++ backend/static/auth/authStyles.css
@@ -97,3 +97,15 @@ button:hover {
     color: #007bff;
     text-decoration: none;
 }
+.fade-in {
+    animation: fadeIn 0.5s ease-in-out;
+}
+
+@keyframes fadeIn {
+    from {
+        opacity: 0;
+    }
+    to {
+        opacity: 1;
+    }
+}
