+++ backend/static/auth/login.html
@@ -33,6 +33,6 @@
                 <p>Want to try the FaceID instead? <a href="/recognize">FaceID</a></p>
             </div>
         </div>
-        <script src="auth/login.js" type="text/javascript"></script>
+        <script src="auth/login.js" type="module"></script>
     </body>
 </html>
