+++ backend/static/auth/login.js
@@ -1,19 +1,8 @@
 const loginFormEl = document.querySelector("#loginForm");
-const errorMessageEl = document.querySelector("#error-message");
-const successMessageEl = document.querySelector("#success-message");
 const usernameInputEl = document.querySelector("#username");
 const passwordInputEl = document.querySelector("#password");
-function writeSuccessMessage(message) {
-    errorMessageEl.style.display = "none";
-    successMessageEl.textContent = message;
-    successMessageEl.style.display = "block";
-}
-function writeErrorMessage(message){
-        successMessageEl.style.display = "none";
-        errorMessageEl.textContent = message;
-        errorMessageEl.style.display = "block";
 
-}
+import {writeErrorMessage, writeSuccessMessage} from "./messageDisplay.js";
 window.onload = function () {
     const urlParams = new URLSearchParams(window.location.search);
     const error = urlParams.get("error");
@@ -21,7 +10,6 @@ window.onload = function () {
         writeErrorMessage(error)
     }
     const success = urlParams.get("success");
-    console.log(success);
     if (success) {
         writeSuccessMessage(success)
     }
