+++ backend/static/auth/rec.css
@@ -71,3 +71,44 @@ button:hover {
     color: #007bff;
     text-decoration: none;
 }
+.error {
+    color: #dc143c;
+    display: none;
+    background-color: #f8d7da;
+    border: 1px solid #dc143c;
+    border-radius: 5px;
+    text-align: center;
+    width: 100%;
+    box-sizing: border-box;
+    font-size: 16px;
+    font-weight: bold;
+    /* margin-left: 25%; */
+    margin-bottom: 15px;
+}
+.success {
+    color:green;
+    display: none;
+    background-color:#d4edda ;
+    border: 1px solid #1dab51;
+    border-radius: 5px;
+    text-align: center;
+    width: 100%;
+    box-sizing: border-box;
+    font-size: 16px;
+    font-weight: bold;
+    /* margin-left: 25%; */
+    margin-bottom: 15px;
+
+}
+.fade-in {
+    animation: fadeIn 0.5s ease-in-out;
+}
+
+@keyframes fadeIn {
+    from {
+        opacity: 0;
+    }
+    to {
+        opacity: 1;
+    }
+}
